+++ gorp/docs/source/conf.py
@@ -22,7 +22,7 @@ copyright = '2021, Mark J Olson'
 author = 'Mark J Olson'
 
 # The full version, including alpha/beta/rc tags
-release = '0.4.1'
+release = '1.0.0'
 
 
 # -- General configuration ---------------------------------------------------